Split index creation off for easier modification
